pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pyperfect"
version = "0.1.0"
description = "A modern, high-performance Python framework showcasing best practices in 2025."
authors = [{name = "Thomas F McGeehan V", email = "mcgeehan@gmail.com"}]
license = {file = "LICENSE"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "grpcio",
    "grpcio-tools",
    "ibis-framework[duckdb]",
    "pydantic",
    "loguru",
    "pydantic",
    "pytest",
    "httpx",
    "py4j"
]

[project.optional-dependencies]
dev = ["black", "flake8", "mypy", "pytest", "pytest-cov"]
docs = ["mkdocs", "mkdocs-material"]

[tool.setuptools]
packages = ["pyperfect"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--strict-markers --disable-warnings"
testpaths = ["tests"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
strict = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.isort]
profile = "black"
